說明文字

+++ sql/mysql.sql
@@ -7,7 +7,7 @@ CREATE TABLE `exam` (
   `upload_mode` enum('1','0') NOT NULL DEFAULT '1',
   `ext_file` varchar(255) NOT NULL DEFAULT '',
   `note` text NOT NULL,
-  `uid` smallint(5) unsigned NOT NULL DEFAULT '0',
+  `uid` int(10) unsigned NOT NULL DEFAULT '0',
   `open_show` enum('1','0') NOT NULL DEFAULT '1',
   `create_date` date NOT NULL,
   PRIMARY KEY (`assn`)
